chore: create reward helpers file
