+++ src/utils/webhook.js
@@ -15,6 +15,7 @@ import { addIdentitySegment, addTrackSegment } from './segment.js';
 import axios from 'axios';
 import 'dotenv/config';
 import { sendTelegramMessage } from './telegram.js';
+import { reward_helpers } from './rewardHelpers.js'
 
 /**
  * Handles the signup reward for a user.
@@ -81,7 +82,7 @@ export async function handleSignUpReward(db, params) {
           `[${params.eventId}] was stopped due to too long treatment duration (> 10 min).`
         );
 
-        await updateReferralReward(
+        await reward_helpers.updateReferralReward(
           db,
           {
             userTelegramID: params.userTelegramID,
@@ -110,7 +111,7 @@ export async function handleSignUpReward(db, params) {
         }
       } else {
         // Update the reward record to mark it as successful
-        await updateReferralReward(
+        await reward_helpers.updateReferralReward(
           db,
           {
             userTelegramID: params.userTelegramID,
@@ -151,7 +152,7 @@ export async function handleSignUpReward(db, params) {
       const dateAdded = new Date();
 
       // Update the reward record to mark it as successful
-      await updateReferralReward(
+      await reward_helpers.updateReferralReward(
         db,
         {
           userTelegramID: params.userTelegramID,
@@ -201,7 +202,7 @@ export async function handleSignUpReward(db, params) {
     }
 
     if (txReward.data.userOpHash) {
-      await updateReferralReward(
+      await reward_helpers.updateReferralReward(
         db,
         {
           userTelegramID: params.userTelegramID,
@@ -335,7 +336,7 @@ export async function handleReferralReward(db, params) {
           `[${params.eventId}] was stopped due to too long treatment duration (> 10 min).`
         );
 
-        await updateReferralReward(
+        await reward_helpers.updateReferralReward(
           db,
           {
               eventId: params.eventId,
@@ -367,7 +368,7 @@ export async function handleReferralReward(db, params) {
         }
       } else {
         // Update the reward record to mark it as successful
-        await updateReferralReward(
+        await reward_helpers.updateReferralReward(
           db,
           {
             eventId: params.eventId,
@@ -411,7 +412,7 @@ export async function handleReferralReward(db, params) {
     if (txReward.data.txHash) {
       const dateAdded = new Date();
 
-      await updateReferralReward(
+      await reward_helpers.updateReferralReward(
         db,
         {
           eventId: params.eventId,
@@ -462,7 +463,7 @@ export async function handleReferralReward(db, params) {
     }
 
     if (txReward.data.userOpHash) {
-      await updateReferralReward(
+      await reward_helpers.updateReferralReward(
         db,
         {
           eventId: params.eventId,
@@ -576,7 +577,7 @@ export async function handleLinkReward(
           `[${eventId}] was stopped due to too long treatment duration (> 10 min).`
         );
 
-        await updateReferralReward(
+        await reward_helpers.updateReferralReward(
           db,
           {
             userTelegramID: referentUserTelegramID,
@@ -609,7 +610,7 @@ export async function handleLinkReward(
         }
       } else {
         // Update the reward record to mark it as successful
-        await updateReferralReward(
+        await reward_helpers.updateReferralReward(
           db,
           {
             eventId: eventId,
@@ -650,7 +651,7 @@ export async function handleLinkReward(
       const dateAdded = new Date();
 
       // Add the reward to the "rewards" collection
-      await updateReferralReward(
+      await reward_helpers.updateReferralReward(
         db,
         {
           userTelegramID: referentUserTelegramID,
@@ -695,7 +696,7 @@ export async function handleLinkReward(
     }
 
     if (txReward.data.userOpHash) {
-      await updateReferralReward(
+      await reward_helpers.updateReferralReward(
         db,
         {
           userTelegramID: referentUserTelegramID,
@@ -1169,50 +1170,6 @@ export async function handleNewTransaction(params) {
   }
 
   return false;
-}
-
-async function updateReferralReward(
-  db,
-  params
-) {
-  const REWARDS_COLLECTION = 'rewards';
-
-  try {
-    const $set = {};
-
-    if (params.eventId) $set.eventId = params.eventId;
-    if (params.userTelegramID) $set.userTelegramID = params.userTelegramID;
-    if (params.responsePath) $set.responsePath = params.responsePath;
-    if (params.walletAddress) $set.walletAddress = params.walletAddress;
-    if (params.reason) $set.reason = params.reason;
-    if (params.userHandle) $set.userHandle = params.userHandle;
-    if (params.userName) $set.userName = params.userName;
-    if (params.amount) $set.amount = params.amount;
-    if (params.message) $set.message = params.message;
-    if (params.transactionHash) $set.transactionHash = params.transactionHash;
-    if (params.status) $set.status = params.status;
-    if (params.dateAdded) $set.dateAdded = params.dateAdded;
-    if (params.userOpHash) $set.userOpHash = params.userOpHash;
-    if (params.newUserAddress) $set.newUserAddress = params.newUserAddress;
-    if (params.parentTransactionHash) $set.parentTransactionHash = params.parentTransactionHash;
-    if (params.sponsoredUserTelegramID) $set.sponsoredUserTelegramID = params.sponsoredUserTelegramID;
-
-    const result = await db.collection(REWARDS_COLLECTION).updateOne(
-      {
-        userTelegramID: params.userTelegramID,
-        eventId: params.eventId,
-        reason: params.reason,
-      },
-      {
-        $set
-      },
-      { upsert: true }
-    );
-
-    return result;
-  } catch (error) {
-    throw error;
-  }
 }
 
 export const webhook_utils = {
